Rearrange Codebase.
